requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
autopep8==2.3.2
aws-requests-auth==0.4.3
beautifulsoup4==4.14.2
bedrock-agentcore==1.0.3
bedrock-agentcore-starter-toolkit==0.1.26
boto3==1.40.55
botocore==1.40.55
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
decorator==5.2.1
dill==0.4.0
docstring_parser==0.17.0
frozenlist==1.8.0
h11==0.16.0
halo==0.0.31
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
log-symbols==0.0.14
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mcp==1.18.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.2.6
ollama==0.6.0
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-python==4.12.0.88
opentelemetry-api==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-threading==0.59b0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pathable==0.4.4
pillow==11.3.0
platformdirs==4.5.0
prance==25.4.8.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.4.1
py-openapi-schema-to-json-schema==0.0.3
pycodestyle==2.14.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
questionary==2.1.1
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rich==14.2.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
s3transfer==0.14.0
scenedetect==0.6.7.1
shellingham==1.5.4
six==1.17.0
slack_bolt==1.26.0
slack_sdk==3.37.0
sniffio==1.3.1
soupsieve==2.8
spinners==0.0.24
sse-starlette==3.0.2
starlette==0.48.0
strands-agents==1.13.0
strands-agents-builder==0.1.10
strands-agents-tools==0.2.12
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
toml==0.10.2
tqdm==4.67.1
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
verboselogs==1.7
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
